# openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MemoIn": {
        "properties": {
          "runId": {
            "title": "Runid",
            "type": "string"
          }
        },
        "required": [
          "runId"
        ],
        "title": "MemoIn",
        "type": "object"
      },
      "MemoOut": {
        "properties": {
          "docxSha256": {
            "title": "Docxsha256",
            "type": "string"
          },
          "docxUrl": {
            "title": "Docxurl",
            "type": "string"
          },
          "htmlUrl": {
            "title": "Htmlurl",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "reviewHead": {
            "default": "",
            "title": "Reviewhead",
            "type": "string"
          },
          "runId": {
            "title": "Runid",
            "type": "string"
          }
        },
        "required": [
          "id",
          "runId",
          "docxUrl",
          "htmlUrl",
          "docxSha256"
        ],
        "title": "MemoOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Contract Workbench",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/memos": {
      "post": {
        "description": "201 with a memo for the run's current review state, or 200 with the one already written under it. A review\nafter a memo makes the next request write a new memo; the earlier one is kept as it was.",
        "operationId": "create_memo_api_memos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MemoIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MemoOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Memo",
        "tags": [
          "memos"
        ]
      }
    },
    "/api/memos/{memo_id}/docx": {
      "get": {
        "operationId": "memo_as_docx_api_memos__memo_id__docx_get",
        "parameters": [
          {
            "in": "path",
            "name": "memo_id",
            "required": true,
            "schema": {
              "title": "Memo Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Memo As Docx",
        "tags": [
          "memos"
        ]
      }
    },
    "/api/memos/{memo_id}/html": {
      "get": {
        "operationId": "memo_as_html_api_memos__memo_id__html_get",
        "parameters": [
          {
            "in": "path",
            "name": "memo_id",
            "required": true,
            "schema": {
              "title": "Memo Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Memo As Html",
        "tags": [
          "memos"
        ]
      }
    },
...
  }
}